--- src/main/java/com/goldstone/saboteur_backend/dtos/gameRoom/GameRoomRequestDto.java
@@ -1,15 +1,0 @@
-package com.goldstone.saboteur_backend.dtos.gameRoom;
-
-import java.util.UUID;
-import lombok.*;
-
-public class GameRoomRequestDto {
-
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Getter
-    @Setter
-    public static class CreateGameRoomRequestDto {
-        private UUID userId;
-    }
-}
