+++ src/main/java/com/goldstone/saboteur_backend/dtos/gameRoom/request/CreateGameRoomRequestDto.java
@@ -1,0 +1,13 @@
+package com.goldstone.saboteur_backend.dtos.gameRoom.request;
+
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class CreateGameRoomRequestDto {
+    private UUID userId;
+}
