+++ src/main/java/com/goldstone/saboteur_backend/dtos/gameRoom/response/CreateGameRoomResponseDto.java
@@ -1,0 +1,20 @@
+package com.goldstone.saboteur_backend.dtos.gameRoom.response;
+
+import com.goldstone.saboteur_backend.domain.GameRoom;
+import com.goldstone.saboteur_backend.dtos.user.response.UserInfoResponseDto;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class CreateGameRoomResponseDto {
+    private String id;
+    private UserInfoResponseDto masterPlayer;
+
+    public static CreateGameRoomResponseDto of(GameRoom gameRoom) {
+        return CreateGameRoomResponseDto.builder()
+                .id(gameRoom.getId().toString())
+                .masterPlayer(UserInfoResponseDto.of(gameRoom.getMaster()))
+                .build();
+    }
+}
