+++ src/main/java/com/goldstone/saboteur_backend/service/gameRoom/GameRoomService.java
@@ -1,3 +1,9 @@
 package com.goldstone.saboteur_backend.service.gameRoom;
 
-public interface GameRoomService {}
+import com.corundumstudio.socketio.SocketIOClient;
+import com.goldstone.saboteur_backend.domain.GameRoom;
+import com.goldstone.saboteur_backend.domain.User;
+
+public interface GameRoomService {
+    void joinUser(SocketIOClient client, GameRoom gameRoom, User user) throws Exception;
+}
