+++ src/main/java/com/goldstone/saboteur_backend/socketIo/SocketIoController.java
@@ -3,8 +3,8 @@ package com.goldstone.saboteur_backend.socketIo;
 import com.corundumstudio.socketio.SocketIOServer;
 import com.corundumstudio.socketio.listener.ConnectListener;
 import com.corundumstudio.socketio.listener.DisconnectListener;
-import com.goldstone.saboteur_backend.dtos.gameRoom.GameRoomRequestDto;
-import com.goldstone.saboteur_backend.service.gameRoom.GameRoomServiceImpl;
+import com.goldstone.saboteur_backend.domain.GameRoom;
+import com.goldstone.saboteur_backend.dtos.gameRoom.request.CreateGameRoomRequestDto;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
@@ -25,8 +25,8 @@ public class SocketIoController {
 
         server.addEventListener(
                 "createGameRoom",
-                GameRoomRequestDto.CreateGameRoomRequestDto.class,
-                (client, data, ackSender) -> GameRoomServiceImpl.createGameRoom(client, data));
+                CreateGameRoomRequestDto.class,
+                (client, data, ackSender) -> GameRoom.createGameRoom(client, data));
     }
 
     public ConnectListener listenConnected() {
