--- src/main/resources/application.properties
@@ -1,18 +1,0 @@
-spring.application.name=saboteur
-
-# jpa
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:13306/mytest
-spring.datasource.username=root
-spring.datasource.password=1234
-
-# redis
-spring.data.redis.host=localhost
-spring.data.redis.port=16379
-
-# socket.io
-socketio.server.hostname=localhost
-socketio.server.port=8081
-
-# disable autoconfigure
-logging.level.org.springframework.boot.autoconfigure=error
